% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofr.R
\name{lofmodel}
\alias{lofmodel}
\title{Make a model object for \code{\link{predict.lofmodel}}.}
\usage{
lofmodel(data, k.min = 1, k.max = 10, knn.args = list(), nparallel = 1)
}
\arguments{
\item{data}{a matrix.}

\item{k.min}{mininum value of k.}

\item{k.max}{maximal value of k.}

\item{knn.args}{optinal arguments for \code{\link{get.knn}}.}

\item{nparallel}{optional, the number of cores to be used for parallel computing.}
}
\value{
a list with class "lodmodel".
}
\description{
Make a model object for \code{\link{predict.lofmodel}}.
}
\examples{
index <- sample(1:150, 100)
dat1 <- iris[index, -5]
dat2 <- iris[setdiff(1:nrow(iris), index), -5]
mod <- lofmodel(dat1, k.min=1, k.max=10)
lof1 <- predict(mod) # local outier factor for dat1
lof2 <- predict(mod, dat2) # local outlier factor for dat2

}
\seealso{
\code{\link{predict.lofmodel}}, \code{\link[FNN]{get.knn}}.
}

