% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofr.R
\name{predict.lofmodel}
\alias{predict.lofmodel}
\title{Calculate local outlier factor.}
\usage{
\method{predict}{lofmodel}(object, newdata = NULL, nparallel = 1)
}
\arguments{
\item{object}{a value of "\code{lofmodel}".}

\item{newdata}{optional, a matrix to calculate LOF. If omitted, the data given at "\code{lofmodel}" is used.}

\item{nparallel}{optional, the number of cores to be used for parallel computing.}
}
\value{
a matrix with the local outlier factor of each observation as rows and each k value as columns
}
\description{
Calculate local outlier factor.
}
\details{
this function does not consider "ties"
}
\examples{
index <- sample(1:150, 100)
dat1 <- iris[index, -5]
dat2 <- iris[setdiff(1:nrow(iris), index), -5]
mod <- lofmodel(dat1, k.min=1, k.max=10)
lof1 <- predict(mod) # local outier factor for dat1
lof2 <- predict(mod, dat2) # local outlier factor for dat2

}
\seealso{
\code{\link{lofmodel}}
}

